Extract temporary personnel menus into a separate class
This change moves code as is. Since all temporary personnel
management menu code is highly isolated, this change simplifies
`MekHQMenuBar` dramatically and simplifies further refactoring
of temp pool management.
